fix lort sina mistake
